running E2E training
