+++ examples/qwen3/qat.yml
@@ -1,9 +1,9 @@
-base_model: Qwen/Qwen3-8B
+base_model: Qwen/Qwen3-0.6B
 # Automatically upload checkpoint and final model to HF
 # hub_model_id: username/custom_model_name
 
 load_in_8bit: false
-load_in_4bit: true
+load_in_4bit: false 
 strict: false
 
 datasets:
@@ -22,7 +22,7 @@ qat:
   activation_dtype: int8
   weight_dtype: int4
   group_size: 32
-  
+
 wandb_project:
 wandb_entity:
 wandb_watch:
@@ -37,31 +37,31 @@ lr_scheduler: cosine
 learning_rate: 0.0002
 
 bf16: auto
-tf32: true
+tf32: false
 
 gradient_checkpointing: true
 gradient_checkpointing_kwargs:
   use_reentrant: false
 resume_from_checkpoint:
 logging_steps: 1
-flash_attention: true
+flash_attention: false
 
 warmup_steps: 10
 evals_per_epoch: 4
 saves_per_epoch: 1
 weight_decay: 0.0
-fsdp:
-  - full_shard
-  - auto_wrap
-fsdp_config:
-  fsdp_version: 2
-  fsdp_offload_params: false
-  fsdp_cpu_ram_efficient_loading: true
-  fsdp_auto_wrap_policy: TRANSFORMER_BASED_WRAP
-  fsdp_transformer_layer_cls_to_wrap: Qwen3TextDecoderLayer
-  fsdp_state_dict_type: SHARDED_STATE_DICT
-  fsdp_sharding_strategy: FULL_SHARD
-  fsdp_reshard_after_forward: true
-  fsdp_activation_checkpointing: true
+# fsdp:
+#   - full_shard
+#   - auto_wrap
+# fsdp_config:
+#   fsdp_version: 2
+#   fsdp_offload_params: false
+#   fsdp_cpu_ram_efficient_loading: true
+#   fsdp_auto_wrap_policy: TRANSFORMER_BASED_WRAP
+#   fsdp_transformer_layer_cls_to_wrap: Qwen3TextDecoderLayer
+#   fsdp_state_dict_type: SHARDED_STATE_DICT
+#   fsdp_sharding_strategy: FULL_SHARD
+#   fsdp_reshard_after_forward: true
+#   fsdp_activation_checkpointing: true
 
 special_tokens:
