+++ src/axolotl/utils/config/__init__.py
@@ -23,19 +23,6 @@ from axolotl.utils.schemas.datasets import DPODataset, KTODataset, SFTDataset
 
 LOG = logging.getLogger("axolotl")
 
-
-class TorchDType(Enum):
-    int1 = torch.int1
-    int2 = torch.int2
-    int3 = torch.int3
-    int4 = torch.int4
-    int5 = torch.int5
-    int6 = torch.int6
-    int7 = torch.int7
-    int8 = torch.int8
-    uint8 = torch.uint8
-
-
 def choose_device(cfg):
     def get_device():
         try:
